Add files via upload
This Python script uses the 'yahooquery' module to help download the historical data for stock 'HSBC Holdings plc'.  This stock is represented with the ticker symbol HSBA.L.  Most UK stocks have a .L when viewed on the Yahoo Finance website.

When the script is run, a CSV file is created.  The CSV file is called 'HSBA.L_yahooquery_stockHistory.csv'.  This CV file can then be used for other scripts if needed.