Add files via upload
This Python script runs a Japanese Candlestick Chart for the stock HSBA.L.  It also shows the purchase date and price of the stock.  The script takes the data from the 'HSBA.L_yahooquery_stockHistory' CSV file.

The script automatically generates the dates for the stock, so there is no need for the user to manually type them each time.

The Purchase Date and Price shows the user how the stock is performing at later dates.