Add files via upload
This image is the output from when running the Python script 'HSBA.L_JapaneseCandlestickChart_OnePurchase'.

The image shows a Japanese Candlestick chart for stock 'HSBA.L'  It represents the stock price movement timeline.

The purchase date of the stock is represented in dashed blue.  The purchase price of the stock is represented in solid blue.