Add files via upload
This image is the output from when running the Python script 'HSBA.L_BarChart_UK__ExDividendDates_USD'.

The image shows a Bar Chart for stock 'HSBA.L'  It represents the dividends of the stock in years 2024 and 2025.  Year 2024 is represented by the colour 'blue' and year 2025 is represented by the colour 'orange'.